requirements.txt
anyio==4.6.2.post1
apify_client==1.8.1
apify_shared==1.1.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
more-itertools==10.5.0
oauthlib==3.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
urllib3==2.2.3